+++ src/form-elements/FormElementBSB.tsx
@@ -162,6 +162,7 @@ function FormElementBSB({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>
         {(((isDirty || displayValidationMessage) &&