+++ src/form-elements/FormElementBarcodeScanner.tsx
@@ -142,6 +142,7 @@ function FormElementBarcodeScanner({
                 value={value}
                 validationMessage={validationMessage}
                 lookupButtonConfig={element.lookupButton}
+                allowLookupOnEmptyValue={element.runLookupOnClear}
               />
             </div>