+++ src/form-elements/FormElementDate.tsx
@@ -123,6 +123,7 @@ function FormElementDate({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>