+++ src/form-elements/FormElementText.tsx
@@ -64,6 +64,7 @@ function FormElementText({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>
         {(isDisplayingValidationMessage || !!element.maxLength) && (