+++ src/form-elements/FormElementTime.tsx
@@ -105,6 +105,7 @@ function FormElementTime({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            allowLookupOnEmptyValue={element.runLookupOnClear}
           />
         </div>